DB Scheema update for document verification
